+++ app/components/organisms/main/hero/hero.module.css
@@ -1,0 +1,33 @@
+@keyframes landing {
+  from {
+    transform: translateY(20px);
+    opacity: 0;
+  }
+  to {
+    transform: translateY(0);
+    opacity: 1;
+  }
+}
+
+@keyframes landing-reverse {
+  from {
+    transform: translateY(-20px);
+    opacity: 0;
+  }
+  to {
+    transform: translateY(0);
+    opacity: 1;
+  }
+}
+
+.heroText {
+  animation: landing-reverse 0.5s ease-in-out forwards 1s;
+}
+
+.heroButton {
+  animation: landing 0.5s ease-in-out forwards 1s;
+}
+
+.heroTitle {
+  animation: landing 0.5s ease-in-out forwards 0.5s;
+}
